requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
APScheduler==3.9.1
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.30.0
colorama==0.4.6
cryptography==38.0.3
dateparser==1.1.3
dnspython==2.2.1
exceptiongroup==1.0.4
Flask==2.2.2
Flask-Cors==3.0.10
frozenlist==1.3.3
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pluggy==1.0.0
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.2
pyparsing==3.0.9
pytest==7.2.0
pytest-env==0.8.1
python-binance==1.0.12
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.1
six==1.16.0
# TA-Lib==0.4.21
# TA-Lib @ file:///C:/Users/Thushalya/Desktop/desktop2/My%20projects/Sem%205%20Project/Project/Data-Explorer-BackEnd/TA_Lib-0.4.24-cp39-cp39-win_amd64.whl
tomli==2.0.1
tzdata==2022.6
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
websocket-client==1.4.1
websockets==10.4
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.10.0
